fusión total entre calendar y drag
