+++ .gitignore
@@ -13,6 +13,7 @@ dist
 dist-ssr
 coverage
 *.local
+db.json
 
 /cypress/videos/
 /cypress/screenshots/
