+++ src/components/Calendar.vue
@@ -13,34 +13,36 @@
                 <button class="date-button" @click="getPrevious"><b>&lt;</b></button>
 
                 <div v-if="selectedComponent == Day">
-                    {{ dayNames[date.subtract(1, 'day').day()] }}, {{ date.date() }} de {{ monthNames[date.month()] }} de {{ date.year() }}
+                    {{ dayNames[date.subtract(1, 'day').day()] }}, {{ date.date() }} de {{ monthNames[date.month()] }}
+                    de {{ date.year() }}
                 </div>
 
                 <div v-if="selectedComponent == Week" class="">
                     {{ date.startOf('week').date() }} de {{ monthNames[date.month()] }} de {{ date.year() }}
                 </div>
-                
+
                 <div v-if="selectedComponent == Month">
                     {{ monthNames[date.month()] }} de {{ date.year() }}
                 </div>
-                
+
                 <button class="date-button" @click="getNext"><b>&gt;</b></button>
             </div>
 
             <div class="option-container">
-                <AddEvent />
+                <AddEvent @SendEvent="addEvent" />
             </div>
         </div>
 
         <div class="body-container">
-            <component :is="selectedComponent" :dayNames="dayNames" :week="week" :month="month" :selectedMonth="selectedMonth" :selectedDay="selectedDay" :currentMonth="currentMonth" />
+            <component :is="selectedComponent" :dayNames="dayNames" :week="week" :month="month"
+                :selectedMonth="selectedMonth" :selectedDay="selectedDay" :currentMonth="currentMonth" :list="items" />
         </div>
     </div>
 </template>
   
 <script setup>
 
-import { ref, shallowRef, watch, onMounted } from "vue";
+import { ref, shallowRef, watch, onMounted, onBeforeMount } from "vue";
 
 import Day from './Day.vue';
 import Week from './Week.vue';
@@ -51,11 +53,10 @@ import AddEvent from "./AddEvent/AddEvent.vue";
 import dayjs from 'dayjs';
 import calendar from 'dayjs/plugin/calendar';
 import updateLocale from 'dayjs/plugin/updateLocale';
+import axios from "axios";
 
 dayjs.extend(calendar);
 dayjs.extend(updateLocale);
-
-//
 
 const localeObject = {
     name: 'es',
@@ -65,8 +66,6 @@ const localeObject = {
 dayjs.locale('es-my-settings', localeObject);
 
 console.log(dayjs().date())
-
-//
 
 const selectedComponent = shallowRef(Day);
 
@@ -113,11 +112,18 @@ const monthNames = [
     "diciembre"
 ]
 
+//variable de los contenedores
+const items = ref([])
+
 onMounted(() => {
     getWeekRange(date.value);
     getMonthRange(date.value);
     selectedMonth.value = (date.value.month());
     selectedDay.value = (date.value.date());
+})
+// actualizar calendario 
+onBeforeMount(() => {
+    loadEvent();
 })
 
 const getWeekRange = (date) => {
@@ -132,9 +138,9 @@ const getMonthRange = (date) => {
     for (let i = 0; i < date.daysInMonth(); i++) {
         month.value.push(date.startOf('month').add(i, 'day'))
     }
-    
+
     //si el mes no comienza en lunes...
-    if(month.value[0].day() != 1){
+    if (month.value[0].day() != 1) {
         let aux = [];
         for (let i = 0; i < month.value[0].subtract(1, 'day').day(); i++) {
             aux.push(month.value[0].startOf('week').add(i, 'day'));
@@ -142,7 +148,7 @@ const getMonthRange = (date) => {
         month.value = [...aux, ...month.value];
     }
     //si el mes no acaba en domingo...
-    if(month.value[month.value.length - 1].day() != 0){
+    if (month.value[month.value.length - 1].day() != 0) {
         let aux = [];
         for (let i = 0; i < 6 - month.value[month.value.length - 1].subtract(1, 'day').day(); i++) {
             aux.push(month.value[month.value.length - 1].add(i + 1, 'day'));
@@ -174,25 +180,54 @@ const getNext = () => {
             break;
 
         case Week:
-        date.value = date.value.add(1, 'week');
+            date.value = date.value.add(1, 'week');
             break;
 
         case Month:
-        date.value = date.value.add(1, 'month');
+            date.value = date.value.add(1, 'month');
             break;
     }
 }
 
-watch(date,(newDate) => {
+watch(date, (newDate) => {
     console.log(newDate.$d);
     getWeekRange(newDate);
     getMonthRange(newDate);
     selectedMonth.value = newDate.month();
 })
 
-watch(selectedComponent,(nc) => {
+watch(selectedComponent, (nc) => {
     console.log(nc);
 })
+// metodos de gestion de eventos 
+
+const loadEvent = async () => {
+
+    try {
+
+        //recojo la información de la base de datos
+        const res = await axios.get(`http://localhost:3000/events/`);
+
+        //recorro la información que hemos recibido y pusheo lo items en la listas según el dia
+        res.data.forEach(element => {
+            items.value.push({ id: element.id, title: element.name, list: changeList(element) });
+        });
+
+    } catch (error) {
+        console.log(error)
+        alert("Problema al traer eventos");
+    }
+}
+
+const addEvent = (event) => {
+    // se carga directamente en el array que corresponda y ya lo vemos por pantalla 
+    items.value.push({ id: event.id, title: event.name, list: changeList(event) });
+
+}
+
+const changeList = (element) => {
+    return parseInt(element.start_date.substring(element.start_date.length - 2))
+}
 
 </script>
 
@@ -202,10 +237,10 @@ watch(selectedComponent,(nc) => {
 }
 
 .header-container {
-    @apply basis-1/12 flex justify-around items-center 
+    @apply basis-1/12 flex justify-around items-center
 }
 
-.select-container{
+.select-container {
     @apply basis-1/4 p-4
 }
 
@@ -213,19 +248,19 @@ watch(selectedComponent,(nc) => {
     @apply px-6 py-4 rounded-full cursor-pointer bg-[#f6f6f6] hover:bg-[#d6d6d6]
 }
 
-option{
+option {
     background-color: #f6f6f6;
 }
 
-.date-container{
+.date-container {
     @apply basis-1/2 flex justify-between items-center p-4 gap-8
 }
 
-.date-button{
+.date-button {
     @apply px-6 py-4 rounded-full hover:bg-[#d6d6d6]
 }
 
-.option-container{
+.option-container {
     @apply basis-1/4 flex justify-end p-4
 }
 
@@ -234,13 +269,12 @@ option{
 }
 
 .body-container::-webkit-scrollbar {
-  display: none;
+    display: none;
 
 }
 
-.body-container  {
-  -ms-overflow-style: none;
-  scrollbar-width: none;
+.body-container {
+    -ms-overflow-style: none;
+    scrollbar-width: none;
 }
-
 </style>
